Move API endpoints from Application to StudentController
