+++ src/main/java/raisetech/StudentManagement5/Application.java
@@ -1,33 +1,13 @@
 package raisetech.StudentManagement5;
 
-import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-import raisetech.StudentManagement5.data.Student;
-import raisetech.StudentManagement5.data.StudentCourse;
 
 @SpringBootApplication
-@RestController
 public class Application {
-
-  @Autowired
-  private StudentRepository repository;
 
   public static void main(String[] args) {
     SpringApplication.run(Application.class, args);
-  }
-
-  @GetMapping("/studentList")
-  public List<Student> getStudentList() {
-    return repository.searchStudent();
-  }
-
-  @GetMapping("/studentCourseList")
-  public List<StudentCourse> getStudentCourseList() {
-    return repository.searchStudentCourse();
   }
 
 }
